all-new workunit size computation logic

+++ webadmin/fitcrackAPI/src/src/api/fitcrack/responseModels.py
@@ -124,7 +124,7 @@ job_short_model = api.model('Job short', {
     'status_type': fields.String(),
     'progress': fields.Float(required=False),
     'time': fields.DateTime(required=False),
-    'workunit_sum_time': fields.Float(),
+    'workunit_sum_time': fields.Integer(),
     'hash_type': fields.String(required=True),
     'time_start': fields.DateTime(required=True),
     'time_end': fields.DateTime(required=True),
